pyproject.toml
[project]
name = "yaclaw"
version = "0.1.0"
description = "Yet another minimal alternative to OpenClaw."
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "apscheduler>=3.11.2",
    "discord-py>=2.6.4",
    "pexpect>=4.9.0",
    "python-dotenv>=1.2.1",
]
